requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
constantly==15.1.0
cryptography==3.2.1
cssselect==1.1.0
Django==3.1.5
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.11.2
dnspython==2.0.0
et-xmlfile==1.0.1
eventlet==0.25.2
greenlet==0.4.17
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.7.0
itemadapter==0.2.0
itemloaders==1.0.4
jdcal==1.4.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.6.1
Markdown==3.3.4
mccabe==0.6.1
monotonic==1.5
mysql-client==0.0.1
mysql-connector-python==8.0.25
numpy==1.19.3
openpyxl==3.0.6
pandas==1.1.4
parsel==1.6.0
pi==0.1.2
Pillow==8.1.0
Protego==0.1.16
protobuf==3.17.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.4
queuelib==1.5.0
requests==2.24.0
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
stem==1.8.0
termcolor==1.1.0
toml==0.10.2
Twisted==20.3.0
urllib3==1.25.11
w3lib==1.22.0
wrapt==1.12.1
xlrd==2.0.1
zope.interface==5.2.0
